Finalizing position of extension
