+++ app/renderer/types/index.ts
@@ -1,6 +1,5 @@
 export * from './_argument.types';
 export * from './_command.types';
-export * from './_editor.types';
 export * from './_theme.types';
 export * from './_redux.types';
 export * from './_update.types';
